mcp/weather_tool/pyproject.toml
[project]
name = "get_weather"
version = "0.1.0"
description = "simple weather tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "requests>=2.32.3",
    "authlib>=1.6.9",   # Indirect; prevents CVE-2026-27962
    "urllib3>=2.6.3",   # Indirect; prevents CVE-2025-66418
    "python-multipart>=0.0.22", # Indirect; prevents CVE-2026-24486
    "fastmcp>=3.2.0",           # Indirect; prevents CVE-2026-32871
    "cryptography>=46.0.5",     # Indirect; prevents CVE-2026-26007
    "starlette>=0.49.1",        # Indirect; prevents CVE-2025-62727
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
]
